fix: determinism, money-math & reliability remediation from reliability audit
Production real-money metagraph hardening. All consensus-output-changing fixes
are intended for a COORDINATED ATOMIC UPGRADE (all L0/L1 nodes together).

Consensus / determinism
- Validator rewards: delete the live /cluster/info HTTP query (non-deterministic
  per-node view -> fork) and derive recipients from the committed per-network
  seedlists (ml0-<net>, dl1-<net>) read as classpath resources, network selected
  via CL_APP_ENV. Conserving tax split (sum(payouts) == tax; empty-layer & tiny-tax
  edges handled; no div-by-zero / PosLong crash).
- Remove Instant.now from the L0 validateData path (kept on L1 ingress) to avoid
  wall-clock divergence across validators.

Money math (no Double anywhere)
- Bounties return Long DAG; collateral multipliers are exact Ratio (105/100, ...)
  applied in BigInt; 10% validator tax is integer (/10); toTokenAmountFormat and
  aggregation use Math.multiplyExact / addExact (fail-fast, never silently wrap).

Validation / safety
- Enforce exactly one proof; bound DOR billedAmount; replay guard via per-device
  lastCheckInHash (Option = None -> backward compatible, dropped from JSON).
- Replace unsafe .get/unsafeFrom/Math.abs(MinValue) in OneTimeRewards, CsvReader,
  BalanceAdjustmentLoader and Utils with typed/overflow-safe handling.

Resilience, logging, CI
- DorApi (data-L1 ingress only): no retry on 4xx, distinct handling of an
  undecodable 2xx, jitter + capped backoff, no full-body logging.
- De-spam logs: per-device/per-team/per-empty-snapshot lines -> DEBUG; one
  aggregate INFO summary per reward cycle.
- Add CI test gate (.github/workflows/ci.yml) on PR/push (JDK 11).

Tests (61 green): DeterminismTest (combine determinism, round-trip, pre-upgrade
backward-compat decode), RewardConservationTest, SeedlistResourcesTest (all
per-network seedlists bundled and parse to valid DAG addresses); updated
BuildRewardsTest (+1 datolite from conservation) and CollateralTest (Ratio).

Backward compatible: only new persisted field is DeviceInfo.lastCheckInHash
(Option = None); old state decodes unchanged and keeps its hash. Reverted a
canonical-encoder idea that would have changed existing state hashes.

+++ metagraph/modules/shared_data/src/main/scala/com/my/dor_metagraph/shared_data/bounties/Bounty.scala
@@ -3,9 +3,11 @@ package com.my.dor_metagraph.shared_data.bounties
 import com.my.dor_metagraph.shared_data.types.Types.DorAPIResponse
 
 trait Bounty {
+  // Reward amount in whole DAG (integer). Converted to datolites downstream via toTokenAmountFormat.
+  // Integer-only so the reward pipeline carries no Double and stays deterministic across nodes.
   def getBountyRewardAmount(
     dorApiResponse: DorAPIResponse,
     epochModulus  : Long
-  ): Double
+  ): Long
 }
